Dynamic template import implemented
