--- packages/server/src/templates/projects/index.ts
@@ -1,2 +1,0 @@
-export { passwordReset } from './auth/passwordReset';
-export { sample } from './sample/sample';
