+++ packages/server/src/templates/templates.controller.ts
@@ -1,13 +1,12 @@
-import { Controller, Get } from '@nestjs/common';
+import { Body, Controller, Param, Post } from '@nestjs/common';
 import { TemplatesService } from './templates.service';
 
 @Controller('templates')
 export class TemplatesController {
-  constructor(private readonly templateService: TemplatesService) {}
+  constructor(private readonly templateService: TemplatesService) { }
 
-  @Get('/')
-  getTemplate(): any {
-    const templateName = 'passwordReset';
-    return this.templateService.getTemplate(templateName);
+  @Post('/preview/:templateName')
+  postTemplate(@Param('templateName') templateName: string, @Body() body: any): any {
+    return this.templateService.getTemplate(templateName, body);
   }
 }
